refactor(user): simplify user store and enhance user fetching logic
